fix/feat(peg.py): Fix Not predicate and handle Escape chars in lexer
The Not predicate now returns True and evalSequence will have a
special handler to catch that and continue the matching if not
"successfully failed" to match the current char. IOW, sequences like
this will now work: `(!'"' .)`

The escape chars were needed to properly differentiate line breaks
from literal '\n' chars for example.