test/fix(peg/1): Test rep.{1,2} & Commit's operand can be negative
The operand for the Commit instruction needs to be negative in order
to implement the semantics for the Star (*) operator. The is a 14bit
field that is read into a u16 since the first 2bits of the new
variable are certainly incomplete, a cheap way of getting negative
numbers to work is to use a u8 for now and throw away those other
6bits. We may do better later to be able to take advantage of those
other bits and have farther jumps but that's gonna take a while til it
becomes a problem. Bounds check should probably happen earlier though.