nesjs: CPU6502 handle LSR, ASL, ROL, ROR & RTI
Had to add the Accumulator addressing mode and handle that in the
appropriate places. Also had to deal with wrong data in the opcode
table which is such a drag to debug.

The most annoying part here though was that the stack wasn't being
written to the right place. We're off by 0x100 :D