nesjs: PPU2c02: A whole lot of random stuff
Awful commit gathering a whole lot of attempts to wrap my head around
the flow of the PPU. Mostly, I got that it pulses 3x faster than the
CPU and most of what the CPU does is to control the PPU via CTRL/MASK
or PPUSCROLL, load some data either via PPUADDR+PPUDATA or via OMADMA.

Something that isn't clearly described in code, but that's rather
common in how NES games work is that the PPU won't stabilize for some
time and nesdev has recommended waiting for two vblanks before
starting to use the PPU.  The first few instructions executed by a
game will in fact disable the PPU setting PPUCTRL & PPUMASK both to 0
and only set them back up after the two vBlanks. This is a point where
the emulation won't be accurate as this implementation of PPU will be
ready to go as soon as it's instantiated and plugged into a bus
performing reads & writes. That doesn't seem to produce any negative
side effects as far as I can tell though.

The PPU then produces one dot per cycle, 341 dots form a scanline, and
261 scanlines form a frame.  Dots and Scanlines are directly related
to respectively the X and Y coordinates of the screen.  The cycles and
scanlines that go over the space that is rendered in the screen are
used for coordinating with the CPU (via NMI during vBlank for
example).

Another important fact is that the PPU works as a pipeline,
intertwining memory fetching, with signaling and pixel assembling
tasks.  I haven't gotten to how sprites work yet, but here's what I
understand about the background:

Memory fetching happens one byte at a time for 8 cycles, that collects
enough information to assemble a 1x8 of the tile. First it's the name
table byte, then the attribute table, then the background MSB & LSB.

That's as far as I got here. The shifting of the background &
attribute registers isn't there yet though.